+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod asm;
 pub mod ls;
 pub mod parser;
 pub mod sem;
+pub mod semantic;
 pub mod syntax;
 
 mod util;
